+++ util/termutil/is_tty.go
@@ -4,7 +4,11 @@ import "os"
 
 // IsTTY returns true if a terminal is detected
 func IsTTY() bool {
-	if fileInfo, _ := os.Stdout.Stat(); (fileInfo.Mode() & os.ModeCharDevice) != 0 {
+	return isFileDescriptorTTY(os.Stdin) && isFileDescriptorTTY(os.Stdout)
+}
+
+func isFileDescriptorTTY(fd *os.File) bool {
+	if fileInfo, _ := fd.Stat(); (fileInfo.Mode() & os.ModeCharDevice) != 0 {
 		return true
 	}
 	return false